Add methods to coupons controller
I added the index action with the ability to access all instances of the Coupon class so it can list all the coupons in the index view page. I added the create action to take in the params from the user-submitted 'Create New Coupon' form and persist that information as a new record in the database while also creating a new Coupon object. I also set a private method to find a coupon by its ID.

I think the show action will still work properly even though I didn't specifically declare it in my controller because of **Rails magic**. :P We'll find out.